Avoid debug logging related memory allocations
Currently, BenchmarkUDPEcho (an end-to-end test including
client, proxy, and target) uses 118 allocations (6102 bytes)
for each round trip packet.  Of those, approximately 36 (2120 bytes) are
used by the proxy upstream, and 25 (1887 bytes) are used by the proxy
downstream.  (The remainder are in the client and target.)

This change reduces the number of allocations to 104 (5827 bytes), with
26 (1933 bytes) upstream and 22 (1808 bytes) downstream, for 20% fewer
allocations and 7% less memory allocated.

Note that these allocations were all for temporary objects, so the
amount of memory "in use" is not expected to change.