Optimize NAT for DNS
This change closes single-use DNS sockets as soon as a response is forwarded.

If a client sends multiple queries before the first response (as could
happen if the client is multiplexing queries onto a single socket, or
a packet is lost upstream), the socket stays open but with a reduced NAT
timeout.

This change also adds unit tests for the NAT map.