Avoid creating NAT mappings for rejected packets
Currently, each valid shadowsocks packet causes a NAT mapping to be
created (if there isn't a matching one already), even if that packet is
rejected due to a malformed or disallowed destination address. This is
a regression due to #83.  The regression results in a significant memory
leak, because unused mappings are never cleaned up.