🪴 Shadowsocks 2022 UDP protocol change
Change: Server messages are now assigned their own server session IDs, and no longer reuse client session IDs. A new `client session ID` field is added to server message header, between timestamp and padding length.

Motivation: UDP sessions are initiated by clients. When a server restarts during a UDP session, the session state on the server, notably the packet ID, is lost. When the restarted server responds to that session, the packet ID starts again at 0 or 1 << 63. These packets will be treated as replay by the client. For 2022-blake3-aes-256-gcm, the situation is even worse. The separate headers now repeat the same pattern as before, and the AEAD part is effectively doing nonce reuse. This method is completely broken.

This change mitigates the issue by using separate session IDs for client and server messages in a session. The client initiates sessions, and maintains the changing server sessions.

In this implementation, the client maintains two server sessions: old and current. When it receives a server message, the server session ID is checked against the current server session. If the server session ID is different from the current one:

- If the server session ID equals the old session, then the packet is processed using the old session state, and the last seen time of the old session is updated.
- If the client has never seen the server session ID, and the old session's last seen time is less than 60 seconds ago, then the packet is dropped, and an error is returned, indicating that the server session is changing too fast for the client to handle.
- If the client has never seen the server session ID, but the old session's last seen time is over 60 seconds ago, then the new session is accepted after packet validation. The current session is now the old session.

This is the recommended way of handling changing server sessions. It simplifies session management, leaves no room for an adversary to replay packets from the last session, and properly handles out-of-order packets from the old session.